+++ contracts/interfaces/ICToken.sol
@@ -3,9 +3,21 @@
 pragma solidity ^0.8.0;
 
 interface ICToken {
-    function borrow(uint borrowAmount) external returns (uint);
-    function repayBorrow(uint repayAmount) external returns (uint);
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
+
     function underlying() external view returns (address);
-    function getAccountSnapshot(address account) external view returns (uint, uint, uint, uint);
+
+    function getAccountSnapshot(address account)
+        external
+        view
+        returns (
+            uint256,
+            uint256,
+            uint256,
+            uint256
+        );
+
     function comptroller() external view returns (address);
 }
