+++ contracts/interfaces/IConverter.sol
@@ -3,7 +3,9 @@
 pragma solidity ^0.8.0;
 
 interface IConverter {
-    function convert(uint amount) external;
+    function convert(uint256 amount) external;
+
     function source() external view returns (address);
+
     function destination() external view returns (address);
 }
