+++ contracts/interfaces/IPriceOracle.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.0;
 
 interface IPriceOracle {
-    function getUnderlyingPrice(address cToken) external view returns (uint);
+    function getUnderlyingPrice(address cToken) external view returns (uint256);
 }
